requirements.txt
# Core RL and Environment
cyberbattlesim @ git+https://github.com/microsoft/cyberbattlesim.git
gymnasium>=0.29.0
stable-baselines3>=2.0.0
optuna>=3.0.0

# Deep Learning Backend
torch>=2.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Notebooks & Development
jupyter>=1.0.0
ipython>=8.0.0